% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningParameterSelection.R
\name{AIC_select}
\alias{AIC_select}
\title{Tuning parameter selection by AIC}
\usage{
AIC_select(mat.list.t.gau, res.S, m, lam1 = 0.1, lam2 = 0.1, returnJGL = F)
}
\arguments{
\item{mat.list.t.gau}{the list of matrices, which contains the Gaussian distributed counts}

\item{lam1}{numeric, Tuning parameter for the sparsity}

\item{lam2}{numeric, Tuning parameter for the similarities across the conditions}

\item{returnJGL}{logical, whether return the JGL model results. default is F.}
}
\value{
a list of 1) a vector of tuning parameter values and the AIC value; 2) the JGL result object
}
\description{
Tuning parameter selection by AIC
}
