% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mod.R
\name{cov_mod}
\alias{cov_mod}
\title{Function for modifying the covariance matrix to positive definiteness}
\usage{
cov_mod(Y, smin = 0.1, m = NULL)
}
\arguments{
\item{Y}{the list of matrices of scRNA-seq count data}

\item{smin}{a small positive number added to prevent the estimation of library size from being zero}

\item{m}{a replacement value for -Inf when all elements of the covariance matrix for moment estimation are -Inf}
}
\value{
a list of covariance matrices by moment method and positive definite modification
}
\description{
Function for modifying the covariance matrix to positive definiteness
}
