% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mod.R
\name{cov_q}
\alias{cov_q}
\title{Function for calculating the covariance matrix by moment method}
\usage{
cov_q(Y, Smin = 0.1)
}
\arguments{
\item{Y}{the list of matrices of scRNA-seq count data}

\item{Smin}{a small positive number added to prevent the estimation of library size from being zero}
}
\value{
a list of covariance matrices by moment method
}
\description{
Function for calculating the covariance matrix by moment method
}
