% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningParameterSelection.R
\name{getJGLTuningParamResult}
\alias{getJGLTuningParamResult}
\title{Derive the Joint Graphical Lasso estimation result by AIC tuning parameter selection}
\usage{
getJGLTuningParamResult(GauList, res.S, m, l1vec = NULL, l2vec = NULL)
}
\arguments{
\item{GauList}{a list of gaussian distributed matrices}

\item{l1vec}{a vector of candidate values for lambda1, if NULL, the values will be 1:5/20}

\item{l2vec}{a vector of candidate values for lambda2, if NULL, the values will be 1:5/50}
}
\value{
AIC.vec: a table of AIC values ; jgl.res: the JGL result object
}
\description{
Derive the Joint Graphical Lasso estimation result by AIC tuning parameter selection
}
